requirements.txt
numpy
python-dateutil
xmltodict
lxml
pyproj
opencv-python-headless
shapely
scipy
cython
rasterio>=1.3
ortools
requests
typing-extensions
